Update C program to perform all arithmetic operations.c

+++ C Programming Assignments/Codeforwin/Basic C/C program to perform all arithmetic operations.c
@@ -10,8 +10,9 @@ int main()
   printf("Sum = %d\n",num1+num2);
   printf("Difference = %d\n",num1-num2);
   printf("Product = %d\n",num1*num2);
-  printf("Quotient = %d\n",num1/num2);
+  printf("Quotient = %f\n",(float)num1/num2);
   printf("Modulus = %d\n",num1%num2);
    
     return 0;
 }
+
